Adds confirmation prompt before deletion of entries

+++ model.php
@@ -432,7 +432,7 @@
 				$list.="<tr><td>".$date."</td>
 					<td>".$title."</td>
 					<td><a href=\"entries_handler.php?action=view&id=".$e_id."\">View</a>
-					<a href=\"entries_handler.php?action=delete&id=".$e_id."\">Delete</a></td></tr>";
+					<a href=\"entries_handler.php?action=delete&id=".$e_id."\" onclick=\"return confirm_delete();\">Delete</a></td></tr>";
 				if($counter%10==0)
 					$list.="</span>";
 			}
